Add base plot
