+++ src/config.py
@@ -8,7 +8,7 @@ ASSET = {
         "PERSONS": (
             # (tag / name / full / age (birth) / job / call / info)
             ## main
-            ("sherlock", "シャーロック", "ホーマー,シャーロック", 20,(1,1), "male", "勇者", "me:僕")
+            ("sherlock", "シャーロック", "ホーマー,シャーロック", 20,(1,1), "male", "勇者", "me:僕"),
             ("mary", "メアリー", "", 16,(1,1), "female", "武闘家", "me:ウチ"),
             ("wilson", "ウィルソン", "ハドル,ウィルソン", 40,(1,1), "元神官", "me:私"),
             ("lime", "ライム", "", 19,(1,1), "female", "騎士（王女）", "me:私"),
